fix potential bug with agent api modification
add some tests for using Agent objects via API
