+++ app/objects/c_agent.py
@@ -99,9 +99,9 @@ class Agent(BaseObject):
     async def gui_modification(self, **kwargs):
         self.update('group', kwargs.get('group'))
         self.update('trusted', kwargs.get('trusted'))
-        self.update('sleep_min', int(kwargs.get('sleep_min')))
-        self.update('sleep_max', int(kwargs.get('sleep_max')))
-        self.update('watchdog', int(kwargs.get('watchdog')))
+        self.update('sleep_min', int(kwargs['sleep_min']) if 'sleep_min' in kwargs else None)
+        self.update('sleep_max', int(kwargs['sleep_max']) if 'sleep_max' in kwargs else None)
+        self.update('watchdog', int(kwargs['watchdog']) if 'watchdog' in kwargs else None)
 
     async def kill(self):
         self.update('watchdog', 1)
